sn-users/requirements.txt
bcrypt==4.3.0
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
psycopg2-binary==2.9.10
setuptools==75.8.2
